+++ CHANGELOG.rst
@@ -11,6 +11,11 @@ Features
   :py:func:`~xskillscore.me`. (:issue:`202`, :pr:`200`)
   `Andrew Huang`_
 
+Bug Fixes
+~~~~~~~~~
+- :py:func:`~xskillscore.sign_test` now works for ``xr.Dataset`` inputs.
+  (:issue:`198`, :pr:`199`) `Aaron Spring`_
+
 Internal Changes
 ~~~~~~~~~~~~~~~~
 - Added Python 3.7 and Python 3.8 to the CI. Use the latest version of Python 3